[Realizado alterações no Card]
