+++ src/components/card/interfaces/ICard.tsx
@@ -1,0 +1,7 @@
+export interface ICard {
+  id: string;
+  imageUrl: string;
+  details: string[];
+  title: string;
+  description: string;
+}
